chore|docs|feat(services/header): move out ranges to a separate file; more docs; Stop for Syncer
